third_party/blink/web_tests/external/wpt/css/css-values/minmax-number-serialize-expected.txt

This is a testharness.js-based test.
FAIL 'min(.1)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'min(.1)' and 'calc(0.1)' should serialize the same in specified values. expected "calc(0.1)" but got "min(0.1)"
FAIL 'scale(min(.1))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(min(.1))' and 'scale(calc(0.1))' should serialize the same in specified values. expected "scale(calc(0.1))" but got "scale(min(0.1))"
PASS 'min(.1)' as a computed value should serialize as '0.1'.
PASS 'scale(min(.1))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'max(.1)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'max(.1)' and 'calc(0.1)' should serialize the same in specified values. expected "calc(0.1)" but got "max(0.1)"
FAIL 'scale(max(.1))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(max(.1))' and 'scale(calc(0.1))' should serialize the same in specified values. expected "scale(calc(0.1))" but got "scale(max(0.1))"
PASS 'max(.1)' as a computed value should serialize as '0.1'.
PASS 'scale(max(.1))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'min(.1, .2, .3)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'min(.1, .2, .3)' and 'calc(0.1)' should serialize the same in specified values. expected "calc(0.1)" but got "min(0.1, 0.2, 0.3)"
FAIL 'scale(min(.1, .2, .3))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(min(.1, .2, .3))' and 'scale(calc(0.1))' should serialize the same in specified values. expected "scale(calc(0.1))" but got "scale(min(0.1, 0.2, 0.3))"
PASS 'min(.1, .2, .3)' as a computed value should serialize as '0.1'.
PASS 'scale(min(.1, .2, .3))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'max(.1, .2, .3)' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'max(.1, .2, .3)' and 'calc(0.3)' should serialize the same in specified values. expected "calc(0.3)" but got "max(0.1, 0.2, 0.3)"
FAIL 'scale(max(.1, .2, .3))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(max(.1, .2, .3))' and 'scale(calc(0.3))' should serialize the same in specified values. expected "scale(calc(0.3))" but got "scale(max(0.1, 0.2, 0.3))"
PASS 'max(.1, .2, .3)' as a computed value should serialize as '0.3'.
PASS 'scale(max(.1, .2, .3))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'min(.3, .2, .1)' as a specified value should serialize as 'calc(0.1)'. assert_equals: 'min(.3, .2, .1)' and 'calc(0.1)' should serialize the same in specified values. expected "calc(0.1)" but got "min(0.3, 0.2, 0.1)"
FAIL 'scale(min(.3, .2, .1))' as a specified value should serialize as 'scale(calc(0.1))'. assert_equals: 'scale(min(.3, .2, .1))' and 'scale(calc(0.1))' should serialize the same in specified values. expected "scale(calc(0.1))" but got "scale(min(0.3, 0.2, 0.1))"
PASS 'min(.3, .2, .1)' as a computed value should serialize as '0.1'.
PASS 'scale(min(.3, .2, .1))' as a computed value should serialize as 'matrix(0.1, 0, 0, 0.1, 0, 0)'.
FAIL 'max(.3, .2, .1)' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'max(.3, .2, .1)' and 'calc(0.3)' should serialize the same in specified values. expected "calc(0.3)" but got "max(0.3, 0.2, 0.1)"
FAIL 'scale(max(.3, .2, .1))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(max(.3, .2, .1))' and 'scale(calc(0.3))' should serialize the same in specified values. expected "scale(calc(0.3))" but got "scale(max(0.3, 0.2, 0.1))"
PASS 'max(.3, .2, .1)' as a computed value should serialize as '0.3'.
PASS 'scale(max(.3, .2, .1))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'calc(min(.1) + min(.2))' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'calc(min(.1) + min(.2))' and 'calc(0.3)' should serialize the same in specified values. expected "calc(0.3)" but got "calc(min(0.1) + min(0.2))"
FAIL 'scale(calc(min(.1) + min(.2)))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(calc(min(.1) + min(.2)))' and 'scale(calc(0.3))' should serialize the same in specified values. expected "scale(calc(0.3))" but got "scale(calc(min(0.1) + min(0.2)))"
PASS 'calc(min(.1) + min(.2))' as a computed value should serialize as '0.3'.
PASS 'scale(calc(min(.1) + min(.2)))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'calc(max(.1) + max(.2))' as a specified value should serialize as 'calc(0.3)'. assert_equals: 'calc(max(.1) + max(.2))' and 'calc(0.3)' should serialize the same in specified values. expected "calc(0.3)" but got "calc(max(0.1) + max(0.2))"
FAIL 'scale(calc(max(.1) + max(.2)))' as a specified value should serialize as 'scale(calc(0.3))'. assert_equals: 'scale(calc(max(.1) + max(.2)))' and 'scale(calc(0.3))' should serialize the same in specified values. expected "scale(calc(0.3))" but got "scale(calc(max(0.1) + max(0.2)))"
PASS 'calc(max(.1) + max(.2))' as a computed value should serialize as '0.3'.
PASS 'scale(calc(max(.1) + max(.2)))' as a computed value should serialize as 'matrix(0.3, 0, 0, 0.3, 0, 0)'.
FAIL 'calc(.1 + min(.1))' as a specified value should serialize as 'calc(0.2)'. assert_equals: 'calc(.1 + min(.1))' and 'calc(0.2)' should serialize the same in specified values. expected "calc(0.2)" but got "calc(0.1 + min(0.1))"
FAIL 'scale(calc(.1 + min(.1)))' as a specified value should serialize as 'scale(calc(0.2))'. assert_equals: 'scale(calc(.1 + min(.1)))' and 'scale(calc(0.2))' should serialize the same in specified values. expected "scale(calc(0.2))" but got "scale(calc(0.1 + min(0.1)))"
PASS 'calc(.1 + min(.1))' as a computed value should serialize as '0.2'.
PASS 'scale(calc(.1 + min(.1)))' as a computed value should serialize as 'matrix(0.2, 0, 0, 0.2, 0, 0)'.
FAIL 'calc(max(.1) + .1)' as a specified value should serialize as 'calc(0.2)'. assert_equals: 'calc(max(.1) + .1)' and 'calc(0.2)' should serialize the same in specified values. expected "calc(0.2)" but got "calc(max(0.1) + 0.1)"
FAIL 'scale(calc(max(.1) + .1))' as a specified value should serialize as 'scale(calc(0.2))'. assert_equals: 'scale(calc(max(.1) + .1))' and 'scale(calc(0.2))' should serialize the same in specified values. expected "scale(calc(0.2))" but got "scale(calc(max(0.1) + 0.1))"
PASS 'calc(max(.1) + .1)' as a computed value should serialize as '0.2'.
PASS 'scale(calc(max(.1) + .1))' as a computed value should serialize as 'matrix(0.2, 0, 0, 0.2, 0, 0)'.
Harness: the test ran to completion.
